refactor(auth): refactor use cases for auth and test
- Move all the use cases for auth feature to a dedicated folder for architecture refactor.
