+++ src/domain/auth/use-cases/login-user.ts
@@ -1,6 +1,6 @@
-import { LoginDto } from "../auth/dtos/login.dto";
-import { AuthUserEntity } from "../entities/auth-user.entity";
-import { AuthRepository } from "../repositories/auth.repository";
+import { LoginDto } from "@/domain/auth/dtos/login.dto";
+import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
+import { AuthRepository } from "@/domain/repositories/auth.repository";
 
 interface ILoginUserUseCase {
   execute(dto: LoginDto): Promise<AuthUserEntity>;
