+++ src/domain/auth/use-cases/register-user.ts
@@ -1,7 +1,7 @@
-import { RegisterDto } from "../auth/dtos/register.dto";
-import { AuthUserEntity } from "../entities/auth-user.entity";
-import { UserEntity } from "../entities/user.entity";
-import { AuthRepository } from "../repositories/auth.repository";
+import { RegisterDto } from "@/domain/auth/dtos/register.dto";
+import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
+import { UserEntity } from "@/domain/entities/user.entity";
+import { AuthRepository } from "@/domain/repositories/auth.repository";
 
 interface IRegisterUserUseCase {
   execute(dto: RegisterDto): Promise<UserEntity | AuthUserEntity>;
