+++ src/presentation/controller/__tests__/controller.test.ts
@@ -5,12 +5,12 @@ import { LoginDto } from "@/domain/auth/dtos/login.dto";
 import { UpdateUserDto } from "@/domain/dtos/update-user.dto";
 import { RequestResetPasswordEmailDto } from "@/domain/auth/dtos/request-reset-password-email.dto";
 import { AuthRepository } from "@/domain/repositories/auth.repository";
-import { RegisterUser } from "@/domain/use-cases/register-user";
-import { LoginUser } from "@/domain/use-cases/login-user";
+import { RegisterUser } from "@/domain/auth/use-cases/register-user";
+import { LoginUser } from "@/domain/auth/use-cases/login-user";
 import { UpdateUser } from "@/domain/use-cases/update-user";
 import { UpdateUserPassword } from "@/domain/use-cases/update-user-password";
-import { RequestResetPasswordEmail } from "@/domain/use-cases/request-reset-password-email";
-import { LogoutAuth } from "@/domain/use-cases/logout-user";
+import { RequestResetPasswordEmail } from "@/domain/auth/use-cases/request-reset-password-email";
+import { LogoutAuth } from "@/domain/auth/use-cases/logout-user";
 import { RegisterValidator } from "@/infrastructure/validators/register.validator";
 import { LoginValidator } from "@/infrastructure/validators/login.validator";
 import { UpdateUserValidator } from "@/infrastructure/validators/update-user.validator";
@@ -20,12 +20,12 @@ import { UserEntity } from "@/domain/entities/user.entity";
 import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
 
 // Mock dependencies
-jest.mock("@/domain/use-cases/register-user");
-jest.mock("@/domain/use-cases/login-user");
+jest.mock("@/domain/auth/use-cases/register-user");
+jest.mock("@/domain/auth/use-cases/login-user");
 jest.mock("@/domain/use-cases/update-user");
 jest.mock("@/domain/use-cases/update-user-password");
-jest.mock("@/domain/use-cases/request-reset-password-email");
-jest.mock("@/domain/use-cases/logout-user");
+jest.mock("@/domain/auth/use-cases/request-reset-password-email");
+jest.mock("@/domain/auth/use-cases/logout-user");
 jest.mock("@/infrastructure/validators/register.validator");
 jest.mock("@/infrastructure/validators/login.validator");
 jest.mock("@/infrastructure/validators/update-user.validator");
