+++ src/presentation/controller/controller.ts
@@ -1,7 +1,7 @@
 import { AuthRepository } from "@/domain/repositories/auth.repository";
-import { LoginUser } from "@/domain/use-cases/login-user";
-import { RegisterUser } from "@/domain/use-cases/register-user";
-import { LogoutAuth } from "@/domain/use-cases/logout-user";
+import { LoginUser } from "@/domain/auth/use-cases/login-user";
+import { RegisterUser } from "@/domain/auth/use-cases/register-user";
+import { LogoutAuth } from "@/domain/auth/use-cases/logout-user";
 import { LoginValidator } from "@/infrastructure/validators/login.validator";
 import { RegisterValidator } from "@/infrastructure/validators/register.validator";
 import { LogoutValidator } from "@/infrastructure/validators/logout.validator";
@@ -11,9 +11,10 @@ import { Request, Response, NextFunction } from "express";
 import { UpdateUserValidator } from "@/infrastructure/validators/update-user.validator";
 import { UpdateUser } from "@/domain/use-cases/update-user";
 import { UpdateUserPassword } from "@/domain/use-cases/update-user-password";
-import { RequestResetPasswordEmail } from "@/domain/use-cases/request-reset-password-email";
+
 import { RequestResetPasswordEmailValidator } from "@/infrastructure/validators/request-reset-password-email.validator";
 import { CryptoService } from "@/infrastructure/services/crypto.service";
+import { RequestResetPasswordEmail } from "@/domain/auth/use-cases/request-reset-password-email";
 
 export class AuthController {
   constructor(private readonly datasource: AuthRepository) {}
